divided the server into modules
