+++ .idea/jsDialects.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="JSDialectsMappings">
+    <file url="file://$PROJECT_DIR$/../shri-testing-homework" dialect="ES6" />
+  </component>
+</project>
